INTERMEDIATE COMMIT:
in ED_FIT_CHI2.f90, the function delta_replica(a) needs a temporary
bath. Since the functions set/get_dmft_bath(bath_) are now made so
that they only act dmft_bath defined in VARS_GLOBAL, we need to
replicate their behaviour with a dummy Hloc and V that mimick the
structure of the type effective_bath. This is rather ugly, but maybe not
too much.