INTERMEDIATE COMMIT:
Define Nlat as an input variable rather than a global variable.
This prevents the necessity of ED_INPUT_VARS depending on VARS_GLOBAL
end the consequent non-isolation of the latter from the userspace.
If the driver uses other variables to create the cluster (Nx,Ny ecc.) a
check will be needed inside the driver itself.