INTERMEDIATE COMMIT:
rewritten the format of the bath: it is now of the type

N_lambda_1
N_lambda_...
N_lambda_Nbath
V_1
lambda_1_1...lambda_1_Nlambda
V_...
lambda_...
V_Nbath
lambda_Nbath_1...lambda_Nbath_Nlambda

this makes it simpler to extract the real elements to fit. The array to
be given to the fit (or mixing) routines will be
array_bath(Nbath+1:size(array_bath)).
The old baths have to be converted in the new format.