DRIVER BUGFIX
in the CDMFT driver for the BHZ system,  the blocks of the matrix were
build in a mirror-like way, so that H(1,2) was the element C^+_2C_1,
which is wrong. However, the previously obtained converged baths are
still reasonable, only the lambda components of the Self-Energy have
site indices swapped. This fixes anomalies in the periodization routines