BUGFIX:
fixed the function Hreplica_mask: before, it took a matrix as an input,
and was generally evaluated on Hloc. This does not work well if one uses
a bath parametrization that is more rich than Hloc.

However, this opens a problem: we currently fit only the components for
which Hreplica is nonzero. Since the fit is done on the Weiss field or
on Delta, does this really make sense? Is it a convention that is close
enough? Needs some thinking