Revert latest commit and further fix:
I was aware of the modificiations made on the DMFTtools routine to make
the communicator implicit. I didn't update all the cdmft drivers to
comply with this update.
The old way of running the code (some functions performed by master only
in the driver) did not cause problems up to now, but when the search for
the chemical potential was performed we experienced a failure related to
the broadcast.
Removing the broadcast in the AUX_FUNX function (consistend with the old
implementation) fixed the issue, but it is probably a sub-optimal
solution.
Updated the driver and the auxilliary function to reflect the new
workflow. Needs to be tested on the setup where the problem emerged.