Driver update:
in the kagome driver, use the new ger_custom_observable function to get
the value of n at each step from the green's function integral instead
of the eigenvector expression. This seems to speed up convergence and
provoke less jumps in convergence.

The bath fit is still good and the results are comparable with the
previous case.

To be furtherly tested.