UPDATE:
Modified the way the chemical potential is read in fixed density
calculations: as it was, the file "var.restart" was read at the end of
the loop, where XMU was adjusted. This meant that the first loop was
done with a value of XMU that could be very different from the one
previously obtained. This would kick the bath away from the obtained
solution and make reconverging more lengthy and difficult.

Now the check is made at the read_input stage, after the parameters are
read. If NREAD!=0d0 and the restart file is present, modify XMU
accordingly.

var.restart has been renamed xmu.restart for clarity.