First attempt at an efficient algorithm for Tr_bath[rho_AIM]
Here continues the story begun with commit < https://github.com/QcmPlab/LIB_DMFT_ED/commit/8d7345f9b25647af7caf118c4c0102a80b564dcc > on LIB-DMFT-ED. The development has moved to the CDMFT code for two good reasons:

1. here the need for a full nondiagonal algorithm is more compelling, for we always have hopping amplitudes between the cluster sites, so that we can't expect a diagonal density matrix (while it is usual to have multiorbital single-impurity models without a local hopping structure, allowing for that scenario in the dmft code).

2. the absence of the nonsu2 and superc ed_modes and the "ed_total_ud == true tautology" make for a far less cluttered stage, for a first try.

We have tried here to implement a direct, straightforward implementation of the trace-over-the-bath definition. Namely three nested loops:

do i_cluster = 1,clusterDIM       }---> The cluster-restricted
     do j_cluster = 1,clusterDIM  }---> double loop

          do i_bath = 1,bathDIM   ----> The "tracing loop"

               i = i_cluster [+] i_bath
               j = i_cluster [+] i_bath

               rho(i_cluster,j_cluster) += cvec(i)cvec(j)

Where the [+] is a properly defined composition of indices, and the overall performance should be much better wrt the naive double loop on [1:sectorDIM] since clusterDIM << bathDIM < sectorDIM (actually it should be that sectorDIM = bathDIM*clusterDIM).

Much has been done to achieve this:

[ED_SETUP module]
> build_sector subroutine is now overloaded to an "old" and a "new" implementation, where the new one build two different sector_map objects, one for the cluster and the other for the bath. It poses several challenges for we a) don't know a priori how big the restricted maps need to be, within the given sector, b) we _need_ to avoid storing duplicate entries in these restricted maps, to avoid falling back to full sectorDIM spans, c) so we need to search values within the maps while building them.
> a new search function, for _unordered_ arrays, to accomplish previous point c)

[ED_OBSERVABLES module]
> a rewrite of the full algorithm for the cluster_density_matrix computation, accordingly to the pseudo-code above; The naive implementation with just two loops over the whole sectors is still there in commented blocks (bordered by "!==" lines)
> an implementation of the [+] composition rule, through a state2indices-bjoin-indices2states pipeline that preserves the |ClusterUP,BathUP;ClusterDW,BathDW> ordering, so to fall back to the standard Lin-Gubernatis convention (meaning that we actually DON'T join directly i_cluster and i_bath, but split and reorder the spin components).

--------

Unfortunately, this scheme appears to be *inherently flawed*, for a quite subtle reason: if we build the two maps adding "nonduplicate checks" after the global sector check, we correctly select the all and only allowed configurations for both the cluster and the bath, but we lose the linking between them. We "accept" and store indipendently the cluster and bath configurations, so that we have no guarantee that joining an arbitray couple of them, rebuilds an allowed global state: not all the cluster states go with all the bath states, and we have no way to retain this information if we store the states in separate maps!

A thorough debug session has confirmed that clusterDIM*bathDIM gives always 2*sectorDIM, meaning that we are effectively joining together substates that are incompatible with the given sector.

At the moment I have no idea about how to solve the issue, so that any suggestion is most welcome.

______________________
PS: all of this makes for a lot of problems downstream, of course, like having negative i and j values, NaN terms in the computed rho, etc; it may be that there are other problems scattered around, but I think that the central issue about the nonseparability of the maps is enough to stop and rethink about the whole strategy.