CMake update
DEPRECATED MINIMUM VERSION

CMake latest release (3.22.0) complained that "compatibility with CMake < 2.8.12 will be removed from a future version of CMake", so here we update the minimum version requirement.

It should be reasonable: currently we have 3.15.4 on Ulysses module environment and apt provides 3.10.2[3.16.3] on Ubuntu 18[20].04LTS.

RENAMING THE TARGET

Mapped [dmft_ed -> cdmft_ed] everywhere in CMakeLists; actual building routine is unaffected by this, as it should.