Efficient build of Tr_bath[rho_AIM] through new SPARSE MAP
Introduced a new structure for the Fock space map, that incorporates in one object the information about: (a) the full state, (b) the cluster-impurity part, (c) the bath part, without loosing the information about their relationship.

This has been achieved adding a field to the <sector_map> type, with a custom type <sparse_map> (sp), crafted along the lines of the <sparse_matrix_csr> present in ED_SPARSE_MATRIX module. Nevertheless the subfield <map> of type integer is preserved, and so is its use inside the map_(de)allocate implementations, so that retrocompatibility is guaranteed.

<<< DETAILS >>>

ED_SPARSE_MAP

The field sp contains a sparse data structure that allows storing (in dynamically generated arrays, one for each cluster-impurity state i.imp) the all and only __pairs__ {i.imp; i.bath} that are allowed in the given sector. Storing the bath integers paired with their impurity counterpart is all we needed to avoid treating imp and bath parts as independent (which was at the hearth of the problems within commit 3143e483fb1890135a13ccb850cecdadcf82fd5f ).
Doing so in a _sparse_ data structure furthermore allows to avoid wasting memory, which would happen using a static structure for we do not know beforehand how much allowed bath states are expected for each impurity state.

Finally a comment about the <sp_return_intersection()> method: given the sp object and two cluster states (i.imp, j.imp), it extracts an array with all the bath states that belong to _both_ i.imp and j.imp, so to allow yet another easy "cycle over the bath states" way to implement the trace.

ED_SETUP

<build_sector_new> has been renamed to <build_sector_spin> but despite it still takes in and builds two different maps, they are not anymore relative to the cluster and the bath, respectively. Instead we have a *spin-resolved pair of maps*, and each of them implements the new <sparse_map> building, along the standard list of integers. The separation into spinUP and spinDW maps comes very handy since we want everything always ordered according to the LIN-GUBERNATIS scheme, hence a HC->|impUP,impDW>, HB->|bathUP,bathDW> output would need additional bitwise work to compose the global |impUP,bathUP,impDW,bathDW> state, whenever needed.
The new HUP->{|impUP>;|bathUP>}, HDW->{|impDW>;|bathDW>} has allowed for an implementation of the rho_IMP building procedure that avoids calling any bitwise function, making for a more clear and readable code.

The old <build_sector(isector,H)> is still overloaded and now has been renamed to <build_sector_full> (referring to a unique map containing the *full* states |UP,DW>) and left unchanged: i.e. it does not build the <sparse_map> objects. Nothing prevents it, and doing so should not break all the old calls, for all the references and arguments relative to the <sparse_map> treatment are optional. So we might want to upgrade it too, at some point. (but look down at the final note...)

ED_OBSERVABLES

Most notably I have moved all the code relative to the density matrices outside of <lanc_observables()> , defining a *new subroutine* <density_matrix_impurity()>, located in the same module. The separation is still not clean (to test and debug I needed the densities and doccs to be available on memory, so currently they are carried on at the end of <lanc_observables()> and deallocated only after the rho_imp is computed, printed and checked against the semi-analytical formulas. I will polish all of this in the near future.
There are also a lot of new _debug prints and actions, scattered all around. Many of them will probably be pruned before merging to master.

<<< TESTING and BENCHMARK >>>

All the simplest tests (on cdn_hm_2dsquare model) have been passed:

- Norb=1, Nlat=1 case still gives the correct diagonal rho_IMP, up to machine precision (wrt to semi-analytical formulas).

- Norb=1, Nlat=2 cases (Nx=2 vs Ny=2) give the very same diagonal of the matrix (did not check off-diagonal, yet), and the trace is normalized to essentially 1d0.

- Norb=2, Nlat=1 case gives a diagonal matrix, again normalized to 1d0.

- Timing the building along all the tests we can roughly infer the scaling of the computation:

    Ns    |        T
     4    |      <1ms
     5    |      ~1ms
     6    |      ~4ms
     8    |  10÷250ms (*)

    (*): <=10 for the Nlat=2 case, >150ms for the Norb=2 case, don't really know why there is a difference.

> Globally everything seems to work and the computation should be fast enough to be considered for production runs.

<<< NOTES / TODO >>>

+ More serious testing, on different models and with the guidance of extended semi-analytical benchmarks (up to know we have really checked only the Nimp=1 case) will begin shortly. Scaling to greater Ns should also be inspected.

+ All the machinery needed to implement the <sparse_map> type, subtypes and methods is implemented in a new module ED_SPARSE_MAP, with some significant overlap (e.g. the add_to subroutines) with ED_SPARSE_MATRIX: maybe we should consider merging the two in a unique "ED_SPARSE" module, so to avoid some of the boilerplate.

+ The implementation of <build_sector_spin()> is still not fully compliant to the <ed_total_ud == .false.> case. I tried to maintain the needed formal structure everywhere inside the subroutine (and the compiler did not complain therein) but problems arise when calling the build and accessing to the map in the main body of <density_matrix_impurity()>. Namely, as the loops are coded now, we cannot define the maps as tensors of rank(1): they need instead to be scalars, especially when interacting with <sp_return_intersection()>. Feeding H_sigma(iud) to the routines is not a straightforward solution, since we would then need again to recompose the order of the bits (as it was when the two spin parts were merged in the map). Thus, probably we should build 2*Ns_Ud different maps to really have a clean solution...
> So, currently I just assume instead Ns_Ud = 1 and ignore the issue: this does not affect in any way the density_matrix code, for Ns_Ud is actually hard coded to 1 here.
> Nevertheless we should think about it: solving the issue is needed if we want either to upgrade <build_sector_full()> or get rid of it: that's why it is still untouched and overloaded to the public name <build_sector()>.

--------------
Many thanks to @aamaricci for the outright insightful input about the sparse data structure and for providing a very nice test code to showcase the idea in a most detailed way, without all the clutter related to dealing with the actual sectors of an ED calculation.