Driver update: bath structure fix for HM-2dSquare
Until now an alarming problem has emerged whenever we have tried to increase Nbath to a production value. Even calculations that are totally standard with the correspondent DMFT driver (ed_hm_square) proved problematic here, namely:

- Nlat=1
- Norb=1
- Nbath=6
- Uloc=0
> Stopped execution because of "too many gs", since all the sectors were degenerate.

The issue is that while the DMFT driver relies on a "normal" bath-type, thus having essentialy two families of DoFs in the AIM Hamiltonian --the onsite energies "Ek" and the hybridization tunnelings "Vk"-- the current implementation of the replica bath here defines just one Hsym term, amounting to the intra-replica hopping structure (which, in a sense, is a "dual" of the Vk set).
Thus an homologue of the Ek set is missing, making the convergence VERY FAST and the spin-orbital occupations ALMOST IMMOVABLE from the perfect half-filling condition. The degeneracy problems is inherent to the approach. There is no freedom in the bath to lift the obvious degeneracy of the replicas.

So here we add a diagonal term to the Hsym_basis, representing the perfect analogous of normal-bath's Ek values.
They are initialized to a input-controlled *finite* value (sb_field), uniform over all the diagonal (no crystal-field splitting, of course) and equal for all the replicas (but free to evolve independently through the CDMFT loop, that's the key!).

Some observations:

- If we pass sb_field=0d0 the degeneracy problem remains unsolved. We need to lift the degeneracy from the beginning, since the problem arises at the very first diagonalization.

- if we pass a finite sb_field the degeneracy is readily lifted and the convergence goes fast, but we lose control on the filling of the self-consistent solution.

- I tried several values for sb_field without finding any clear indication of an optimum... small values give solutions very far from half-filling and so do also the very large ones.
  Around 1d0 it goes better but we are still far from an acceptable reproduction of the standard DMFT-ED results.

>> Next step would probably be to try imposing a density goal on the convergence.
   Similarly to how it has been done in the cdn_kagome driver I guess...
   Cfr. commit 3228f5a9b29b7fe152f460979b2b2d1ce1db0a2d