Driver update: fixed-density option for HM-2dSquare
Here we import the "kagome-strategy" < 3228f5a9b29b7fe152f460979b2b2d1ce1db0a2d > to efficiently adjust the chemical potential to enforce a fixed-density goal on the convergence.

We retrieve the site-dependent densities from the G(iw), using tail-corrected FFTs through the <fft_get_density> procedure from DMFTtools. Then average over the cluster and feed the result to <search_chemical_potential>, that compares the value to the target given in the input-file.

+ Convergence is quite fast, even with the additional requirement

- If looking at the standard ed_observables the actual quality of the converged solution seems no different whatsoever from the previous try, i.e. without any control on the density, so that one has still to tune carefully the initial value of sb_field (lambda_sym(1)).

+ Nevertheless if one looks just at the "average FFT-density" everything seems to work like a charm

I find the situation suspicious: are we just implementing a feedback on an object that probably was never problematic, so totally missing the real goal of the trick?

> An "investigative" commit will follow shortly...