Intrigue: why fft_dens is so different to ed_dens? [HM-2dSquare]
Referring to commits < 760dbe8 to 3228f5a > I agree that using the "ed_densities" to control the <search_chemical_potential> routine is quite a nightmare and convergence never seems to arrive.

I ultimately encountered the same problem with the HM-2dSquare driver. Briefly, since when I added a diagonal term in Hsym_basis < cfr. dca11b3 >, if I cannot fix a goal on the density, I end up with solutions that appear as very bad, for that they evade half-filling (with consequences also on the double occupancy, of course). As I said, trying to control the chemical potential using the standard ed_densities seems just not feasible.

If instead I just feed <search_chemical_potential> with the average of the "fft_densities" the problem appears to be promptly gone, as the convergence becomes very fast and *this fft_dens* appears to be fixed quickly and smoothly.

But have you tried to look again at the ed_densities? I don't know of course what happens with the Kagome model, but with the code committed here I get, *in the noninteracting limit* and in the most naive Nlat=Norb=1 case:

-------------
 Loop 1)
 Average ED-density:   1.3860226748510371
 Average FFT-density:  0.99999999999973166

 Loop 2)
 Average ED-density:   1.0497874523238633
 Average FFT-density:  0.99999999999946754

 [...] (ED-dens stays pretty much there...)

 Last Loop)
 Average ED-density:   1.0497875289582670
 Average FFT-density:   1.0000000000002727
 [corresponding docc: 0.275513463989]
-------------
All of this with a quite well-tuned value for lambda_sym(1) (which multiplies the identity). Basically found by intensive trial&error.
For uneducated choices of this value (e.g. 1E-3) I can *reach convergence* with even:
 Average ED-density:  0.80840890464863802     (!!!)
 Average FFT-density:  0.99999999999969158

1. Even the "boxed ones" do not seem to me good ed_occupations, not that I can obtain particularly better results with the correspondent normal-bath DMFT driver, but at least it has not to be so much hyper-tuned. It's the "magic" in replica-bath initialization that always gets me worried.

2. Even without arguing about the quality of the output itself, the final ed_occ values are basically the same as with no fixed-density constrain. Which is no surprise looking at how good the fft_dens appears to converge to the requested nread=1d0. It's just that we are targeting a completely different observable, in terms of accuracy.

> Am I missing something?
> Do you also see such huge deviations in the Kagome case?
> Why are the ed_occupations so inaccurate?

Usually I never cared too much about the accuracy of these specific quantities, but right now this may pose quite a concern for the reduced density matrix story, for that the RDMs are also a direct byproduct of the state_cvec values and accordingly I match the elements of the minimal local reduction up to no more than E-015, when I compute them directly from densities and docc.
So if the ed_occupations are so inaccurate, the same fate I expect for the entanglement spectrum and all related quantities. May this really be the case?

Full disclosure: up to now the problem has not emerged since all the testing has been happening with very small Nbath values, for which very accurate noninteracting and atomic limit values are quite easy to obtain. Things like dens=0.999999958040 and docc=0.249999979020.

_____________

These are questions mostly intended for @DMFTSever and @lcrippa, I guess...