The idea: give each replica its own λˢʸᵐ
The replica Hamiltonian is currently defined as a linear decomposition with respect to a symmetry basis:

Hʳᵉᵖˡⁱᶜᵃ = Σ λˢʸᵐ Hˢʸᵐ

So far, the underlying infrastructure enforces the user to define:

- a unique matricial basis Hˢʸᵐ(:) for all the replicas

- a unique expansion λˢʸᵐ(:), equal for all the replicas, as the initial guess for the bath

While the first point is definitely appropriate (they are called replicas for a reason), the second severely limits the freedom in bath initialization, leading to potentially serious degeneracy problems, as indeed has happened with the Hubbard model on the square lattice.

Hence here we introduce an attempt to solve the issue, by increasing by one the rank of the lambdasym_vector array, so to allow having a different entry for each of the Nbath replicas. More specifically for this model, we would try to mantain a fixed λˢʸᵐ(:,2), which corresponds to the hopping structure of the replicas, and a even linspace for λˢʸᵐ(:,1), which would amount to a p-h symmetric on-site spectrum.[look at lines 96-100]

The appropriate modifications to the solver internals will follow shortly.