Intermediate: discard void imag in ed_print_dm
The type of the density matrices is complex(8) since the eigenspace is complex (here state_cvec are used). But apparently the actual values should always be real (given that in LIB_DMFT_ED we use the state_dvecs...), and so would be the resulting density matrix (as a thermally weighted kronecker product of those vectors).

Most porbably we should just change everything to real type and use state_dvecs, but for now let's just skip printing the imaginary part if it is void (all zeros). All the tests so far have confirmed this condition indeed, but let's stay on the careful side for now.

This would:

1. half the disk footprint of the reduced density matrices (which is kind of relevant for larger clusters)

2. make for a easier IO of these matrices with external tools, for they would be actual square matrices on the .dat file.