Squash-Merge branch 'bath_overhaul' (PR#1) [🚧 PARTIALLY BREAKING 🚧]
What's happening?
=================

The 'symmetry-basis' method to initialize the (replica) bath Hamiltonian is currently defined in terms of the linear decomposition:

Hʳᵉᵖˡⁱᶜᵃ = Σˢʸᵐ( λˢʸᵐ Hˢʸᵐ ),

where Hˢʸᵐ are arrays of the same shape as Hˡᵒᶜ and λˢʸᵐ are the corresponding expansion coefficients.

So far, the user-interface provided by < ed_set_Hreplica > has been expecting the passed λˢʸᵐ array to have (/Nsym/) or (/Nineq,Nsym/) shape, thus enforcing the user to initialize all the replicas with the very same Hamiltonian. While internally
each replica has its own set of coefficients, allowing the dmft-fit to optimize them independently, such a degenerate initial state could easily lead to severe diagonalization problems, as indeed has been observed on the 2d Hubbard model. The issue has been previously addressed with an internal rescaling of _some_ λ-values (those corresponding to a diagonal Hˢʸᵐ term), somewhat hiding the process to the user. 

Here we refresh the < ed_set_Hreplica > interface: all the [Nineq]x[Nbath]x[Nsym] λ-values can be defined explicitly in the driver program and passed all together to the solver, thus enabling full, model-dependent, control of the bath initialization.

On the technical side
=====================

    > single clusters: λˢʸᵐ used to be a (/Nsym/)-shaped array (a vector), now you _should_ pass a (/Nbath,Nsym/) array.

    > inequivalent clusters: λˢʸᵐ used to be a (/Nineq,Nsym/) array, now it **MUST** have (/Nineq,Nbath,Nsym/)  shape.

Please notice the _should_ vs **MUST** contrast. The single-cluster implementation provides a back-compatibility patch to enable the user feed < ed_set_Hreplica > the old way: the solver would just copy the same λ-values on each replica, as it used to, and activate the same legacy internal manipulations; while doing so it would politely inform you that such a functionality is deprecated and might be pull away anytime soon. This is not the case for the real-space interface, whereas supporting the legacy (/Nineq,Nsym/) shape would interfere with the newly introduced (/Nbath,Nsym/) one for the single cluster, with many foreseeable dangers. Please be aware that any attempt to pass a (/Nineq,Nsym/) shape would mess with the procedure overloading and result in unintended behavior; we try to catch the wrong call and print a meaningful error message before stopping, but nothing is assured in the general case: take it as totally unsafe territory.

Updated drivers and tests
=========================

• To soothe the pain of the breaking change I patched the two real-space drivers that I'm aware off: < cdn_bhz_2d_edge > and < cdn_bhz_postprocessing_edge >, so to preserve unchanged their intended behavior (λˢʸᵐ has the correct rank but all replicas are initialized with the same values, manually). They compile but no actual runtime test has been performed.

• I checked that < cdn_bhz_2d > compiles without any change. In this sense the back-compatibility patch has been verified. But again, no runtime test performed, so please be careful.

• All the new functionality has been successfully tested within < cdn_hm_2dsquare > which could then serve as an example of what to do in order to fully exploit the updated interface.