Restart of `bath_overhaul` branch
------------------------------------------------------------------------
This commit updates the branch importing all diffs from master (squash),
and prepares the ground for subsequent updates (a generalization of the
current bath structure).
------------------------------------------------------------------------

WORKPLAN

Here we want to extend our freedom with the bath parametrization, within
a replica-like formulation. Let's consider a cluster of two sites, for a
simple discussion.

The current bath Hamiltonian is a collection of 2x2 matrices (one for
each replica), with e.g. the form:

    ⎡ e -t ⎤
H = ⎢      ⎥
    ⎣-t  e ⎦

where different replicas (H_r) would have different e_r and t_r values.

Each replica H_r hybridizes with the interacting cluster with a single
amplitude, meaning that the i-th site in the cluster and the j-th site
in the r-th replica are connected by a hopping V_ijr = V_r δ_ij.

If we rotate the description to the basis where H_r is diagonal we would
have:

           ⎡ e_r - t_r    0     ⎤
U† H_r U = ⎢                    ⎥
           ⎣     0    e_r + t_r ⎦

which looks as two independent bath sites (per replica) but, given that
the relevant object is actually

[V_r  V_r] H_r [V_r  V_r]† ,

in the rotated basis we would end up with 

                ⎡ e_r - t_r    0     ⎤ ⎡ sqrt(2)V_r ⎤
[sqrt(2)V_r  0] ⎢                    ⎥ ⎢            ⎥
                ⎣     0    e_r + t_r ⎦ ⎣     0      ⎦

which has clearly less freedom than a couple of independent bath sites,
being one hybridization locked to zero. To be clear, this has served us
very well with two site clusters on the single band square lattice. But
addressing bigger clusters is becoming increasingly difficult, leading
to the usual phenomenology for over-constrained optimization landscapes.

Hence we want to allow V_ijr = V_ir δ_ij, in a similar spirit of what is
being built into the single site solver as the bath_type=general update.
> ref. https://github.com/QcmPlab/LIB_DMFT_ED/tree/bath_general
As for LIB_DMFT_ED this would be an optional mode, requested within the
input file.

The example above would then become, in the rotated basis:

 
           ⎡ e_A    0  ⎤ ⎡ V_A ⎤
[V_A  V_B] ⎢           ⎥ ⎢     ⎥
           ⎣  0    e_B ⎦ ⎣ V_B ⎦

with 

e_A = e_r - t_r
e_B = e_r + t_r
V_A = (V_1r + V_2r)/sqrt(2)
V_B = (V_1r - V_2r)/sqrt(2)

achieving the maximal freedom for a two-site (single-orbital) cluster.

For a generic N-site cluster we hope that such a relaxation of the bath
parametrization could give enough freedom as to overcome our issues but
still avoiding the traditional "brute-force" treatment with off-diagonal
hybridizations from all cluster sites to all bath sites, which off-loads
to the optimizer all the burden of keeping the symmetries in the final
bath hamiltonian: here we would off-load just the hybridizations choice,
but maintain the a priori symmetry informed expansion of the Hamiltonian
for the replicas. 

If we'd come to need some control over the constraints on the V_ir vals,
as required by group-theory treatments of the punctual symmetries within
the cluster [PRB 78 115102 (2008) - Koch, Sangiovanni and Gunnarsson]
> ref. https://journals.aps.org/prb/abstract/10.1103/PhysRevB.78.115102
we would just need to expose to the user the initialization of those Vs,
as recently done with the "replica lambdas" in this branch, making this
present update a good candidate for an intermediate approach.

A full encoding of the irreducible representations in the bath structure
would instead call for a symmetry expansion of a general V matrix, akin
to what we do with lambda_sym and H_sym. Let's avoid that for now, but
keep in mind that it can be the definitive (ad hoc) solution for complex
clusters.