Intermediate: v field of effective bath structure becomes a 1d array
For BOTH bath_type='replica' and bath_type='general' we initialize a 1d
array of size Nlat*Nspin*Norb as the "Vs" in the dmft bath structure.
This to keep the implementation the most similar wrt the DMFT-ED code,
where the "general" bath is meant to differentiate the hybridization
amplitudes for all the spin-orbitals (while remaining diagonal on them).
...here in fact we are just adding a spatial degree of freedom (Nlat). 

In both cases the initial values assigned to all these Vs are all equal,
so the key difference would only enter at the fitting step, where the
replica bath would always lock all Vs equal, while the new general bath
would allow them to evolve apart. This implies that {get,set}_dmft_bath
gives a different result for the two bath_types, storing Nbath V values
for the replica case, and Nbath*Nlat*Nspin*Norb values for the new one.
Accordingly get_bath_dimension returns different numbers.

With respect to the current implementation in LIB_DMFT_ED I avoid here
to have different fields for the replica unique scalar "v" and the new
1d-vector "vg", so to minimize the code duplication: everything can be
written in terms of a v[1:Nlat*Nspin*Norb] array and we can easily keep
the "replica" version by enforcing them to be always equal. As we said
above this is taken care of within set_dmft_bath, that now contains an
appropriate "select case (bath_type)" branching.

NOTE: the code currently compiles, but all the actual fitting procedures
      still need to be updated to the new structure (annoyingly enough,
      any error therein would be resolved at runtime only... so the more
      care we can take, the better). Nevertheless a run of the replica
      version has been checked to give basically the same fitted bath,
      after just one loop, as the current master. Hence all the common 
      code that has been rewritten for the new shape of the "v" field,
      has been successfully checked for regressions.

OTHER TODO
------------------------------------------------------------------------
Give more attention to write_dmft_bath() and the read-from-file part in
init_dmft_bath() as they are probably messy right now (even more than
they were already in the master). Ideally we want to converge the I/O
format with the single site code, but we need a public procedure aimed
at converting the old format to the new one, so to preserve the data we
already have (which would be annoying to do, but a necessary chore).
------------------------------------------------------------------------