Intermediate
Here we import all the machinery needed to compile the ED_CHI_SPIN and
ED_CHI_DENS modules, as defined in the single site code (with just few
selected changes related to NONSU2 and SUPERC ed_modes, that here are
not present). Next step would be to extend these implementation to the
actual cluster implementation but perhaps is better to do some testing
on the single site case before. We should also exploit the newly added
procedures* into ED_OBSERVABLES, ED_GF_NORMAL, etc. so to complete the
refactoring and benefit from the inherent code simplification.

*) all the apply_op_xxx procedures, build_op_NS, etc.