Start of branch `orbital_resolved_rdm`
The idea is to generalize the subtracing routine to allow the user to
select in total freedom which orbitals to keep in the reduced system.

This leads to two main challenges:

- trace away orbitals that are sandwiched within the ones we want to
  keep, so care is needed to correctly assign the Fermionic sign
  > thanks @CarlosMejZ for the nice suggestions on the matter

- define a user API to allow selecting the desired orbitals to keep in
  a comfortable ---but generic enough--- way.
  > this we are failing so far, as it is not flexible enough.

So I will proceed to improve the API and hopefully recover both the
missing flexibility and the back-compatibility as an overloading of
the old implementation.

Co-authored-by: Adriano Amaricci <adriano.amaricci@gmail.com>