Test passed on two-site RDMs
Run a single loop on top of a converged point, within the large U Mott
phase of the 2x2 plaquette solution of the 2d Hubbard model (square).

We compare the RDMs by looking at the eigenvalues.

• The two diagonals, i.e. the NNN subsystems, give eigenvalues equal
  up to 1d-3 (remember that sum(eigvals)=1)

• The parallel NN subsystems (the two parallel to x or the two parallel
  to y) give equal eigenvalues up to 1d-5

• The orthogonal NN subsystems (one parallel to x and one to y) give the
  least similar RDMs, with eigenvalues equal only up to 6d-2.

Overall it appears as the 2x2 full cluster RDM might have some numerical
anisotropy, on a noticeable level, albeit not necessarily dramatic.

Note that if the selected sector does not contain the groundstate (as we
did for the first quick comparisons) the effect can be much more evident,
as indeed is expected for excited states.

More testing underway…

————————————————————————————————————————————————————————————————————————

PS. Fixed a silly error on the filename handling for the RDMs
    (missing trim() call on the suffix)