BACKPORT: fix bug on finite chemical potential calculations
While testing on branch orbital_resolved_rdm we found a bug for finite
chemical potential calculations. Namely the chem pot was added to the
bath Hamiltonian (and to the Weiss field). This was a remnant of some
old development trials, where I think the goal was trying to reduce
the number of parameters in the bath (probably the bath was still the
"normal" one). Reference commit: a85ae3a01d4e84f60accf6ffb5287a2490d24792

More info about the testing carried with AA and SG can be found at
> a245e167ddfe09388c00eebd07cf4a9fd9569152 (orbital_resolved_rdm branch)

Here we backport the isolated fix to the master branch.

Co-authored-by: Samuele Giuli <samuele.giuli@gmail.com>
Co-authored-by: Adriano Amaricci <adriano.amaricci@gmail.com>